# Conflicts:
#	api/src/main/javascript/META-INF/resources/myfaces/_impl/xhrCore/_AjaxUtils.js
#	impl/src/main/java/org/apache/myfaces/component/validate/ValidateWholeBeanComponent.java

+++ api/src/main/java/jakarta/faces/convert/BigDecimalConverter.java
@@ -93,7 +93,7 @@ public class BigDecimalConverter
         
         try
         {
-            return ((BigDecimal)value).toString();
+            return value.toString();
         }
         catch (Exception e)
         {
